@@ -16,19 +16,19 @@
     "patch:hosts": "fec patch-etc-hosts",
     "start": "HOT=true fec dev",
     "test": "jest",
-    "postinstall": "rimraf .cache",
+    "postinstall": "ts-patch install && rimraf .cache",
     "verify": "npm-run-all build lint test",
     "prepare": "husky install"
   },
   "dependencies": {
-    "@patternfly/react-core": "^4.276.12",
-    "@patternfly/react-table": "^4.113.4",
-    "@redhat-cloud-services/frontend-components": "^3.11.5",
-    "@redhat-cloud-services/frontend-components-notifications": "^3.2.16",
-    "@redhat-cloud-services/frontend-components-utilities": "^3.7.6",
+    "@patternfly/react-core": "^5.0.0",
+    "@patternfly/react-table": "^5.0.0",
+    "@redhat-cloud-services/frontend-components": "^4.0.1",
+    "@redhat-cloud-services/frontend-components-notifications": "^4.0.2",
+    "@redhat-cloud-services/frontend-components-utilities": "^4.0.2",
     "classnames": "^2.3.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "8.1.2",
     "react-router-dom": "6.15.0",
     "redux": "4.2.1",
@@ -37,18 +37,15 @@
   },
   "devDependencies": {
     "@babel/core": "7.22.11",
-    "@babel/plugin-proposal-object-rest-spread": "7.20.7",
-    "@babel/plugin-transform-runtime": "7.22.15",
     "@babel/preset-env": "7.22.10",
     "@babel/preset-react": "7.22.15",
     "@babel/preset-typescript": "^7.22.11",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
     "@redhat-cloud-services/eslint-config-redhat-cloud-services": "^2.0.2",
-    "@redhat-cloud-services/frontend-components-config": "^5.1.1",
+    "@redhat-cloud-services/frontend-components-config": "^6.0.5",
     "@redhat-cloud-services/tsc-transform-imports": "^1.0.4",
     "@testing-library/jest-dom": "^6.1.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/react": "^14.0.0",
     "@types/react": "17.0.39",
     "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
@@ -56,7 +53,6 @@
     "@typescript-eslint/eslint-plugin": "^6.7.0",
     "@typescript-eslint/parser": "^6.6.0",
     "babel-jest": "29.6.4",
-    "babel-plugin-transform-imports": "^2.0.0",
     "eslint": "8.49.0",
     "husky": "^8.0.3",
     "identity-obj-proxy": "3.0.0",
