@@ -1,7 +1,15 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React, { useEffect } from 'react';
+import { createRoot } from 'react-dom/client';
 import AppEntry from './AppEntry';
 
-const root = document.getElementById('root');
+function EntryWithCallback() {
+  useEffect(() => {
+    root?.setAttribute('data-ouia-safe', 'true');
+  });
 
-ReactDOM.render(<AppEntry />, root, () => root?.setAttribute('data-ouia-safe', 'true'));
+  return <AppEntry />;
+}
+
+const root = document.getElementById('root') as HTMLElement;
+const rootInstance = createRoot(root);
+rootInstance.render(<EntryWithCallback />);
